Kotlin file for adding a restaurant to the mobile app, I added a button in the menu to add a restaurant. The app will ask for permission to locate the phone and read the phone's gallery. The interface has five boxes to fill in (name, rate, GPS location, description, and photo) and one button to upload the recommendation.
All boxes have to be filled in. Otherwise, uploading will be impossible. The cross button X permits returning to the menu interface.

How to fill it in:

Name: The name of the restaurant

Rate: the rate of the restaurant

GPS location: Location of the restaurant (longitude and latitude); if the location of the phone is available, the box will be filled in automatically.

Description: A description or opinion of the restaurant

Icon: Click on it to add a picture of the restaurant from your gallery.

When everything is done, press the upload button.